Revert "renamed Release to ReleaseWorkflow"
This reverts commit 6605ab188c988173db004e74d6fbc010afdfd7c3.
