+++ src/release/index.ts
@@ -1,3 +1,3 @@
-export * from "./release-workflow";
+export * from "./release";
 export * from "./publisher";
 export * from "./release-trigger";
